requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic
pydantic-settings
python-dotenv
langchain
google-generativeai
pytest
pytest-asyncio
httpx